駅指定画面の作成

+++ app/src/main/java/com/example/suzukitakahiro/trainalert/Db/MasterDb/StationDao.java
@@ -33,7 +33,7 @@ public class StationDao {
     /**
      * 都道府県コードに一致する沿線コードを取得する
      */
-    public Loader<Cursor> findByPrefCd(long prefCd) {
+    public Loader<Cursor> findByPrefCd(int prefCd) {
 
         // 取得する情報
         String[] projection = {
